+++ src/console.rs
@@ -15,7 +15,12 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with TRINCI. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::{cheats, client::Client, common::build_transaction, impexp, utils};
+use crate::{
+    cheats,
+    client::Client,
+    common::{build_transaction, FUEL_LIMIT},
+    impexp, utils,
+};
 use trinci_core::{
     crypto::{Hash, KeyPair},
     Transaction,
@@ -30,6 +35,7 @@ const GET_TX: &str = "get-tx";
 const GET_RX: &str = "get-rx";
 const GET_ACCOUNT: &str = "get-acc";
 const GET_BLOCK: &str = "get-blk";
+const GET_TXS: &str = "get-txs";
 const EXPORT_TXS: &str = "export-txs";
 const IMPORT_TXS: &str = "import-txs";
 const CHEATS: &str = "cheats";
@@ -41,7 +47,11 @@ fn help() {
     println!(" * '{} <tkt>': get transaction by ticket", GET_TX);
     println!(" * '{} <tkt>': get receipt by transaction ticket", GET_RX);
     println!(" * '{} <id>': get account by id", GET_ACCOUNT);
-    println!(" * '{} <height>': get block at a given height", GET_BLOCK);
+    println!(
+        " * '{} <height>': get block at a given height (use `max` for the last)",
+        GET_BLOCK
+    );
+    println!(" * '{}': get the number of transactions executed", GET_TXS);
     println!(
         " * '{} <file>': export all transactions into a file",
         EXPORT_TXS
@@ -54,9 +64,12 @@ fn help() {
     println!(" * '{}': exit the application", QUIT);
 }
 
-fn build_transaction_interactive(caller: &KeyPair) -> Transaction {
-    utils::print_unbuf("  Network: ");
-    let network = utils::get_input();
+fn build_transaction_interactive(caller: &KeyPair, config_network: &str) -> Transaction {
+    utils::print_unbuf(&format!("  Network [{}]: ", config_network));
+    let mut network = utils::get_input();
+    if network.is_empty() {
+        network = config_network.to_string();
+    }
 
     utils::print_unbuf("  Target account: ");
     let target = utils::get_input();
@@ -75,8 +88,8 @@ fn build_transaction_interactive(caller: &KeyPair) -> Transaction {
         }
     };
 
-    utils::print_unbuf("  Fuel Limit: ");
-    let fuel_limit = utils::get_input().parse::<u64>().unwrap();
+    utils::print_unbuf(&format!("  Fuel Limit [{}]: ", FUEL_LIMIT));
+    let fuel_limit = utils::get_input().parse::<u64>().unwrap_or(FUEL_LIMIT);
 
     utils::print_unbuf("  Method: ");
     let method = utils::get_input();
@@ -135,8 +148,10 @@ pub fn run(mut client: Client) {
 
         match command {
             PUT_TX => {
-                let tx = build_transaction_interactive(&client.keypair);
-                client.put_transaction_verb(tx);
+                let tx = build_transaction_interactive(&client.keypair, &client.network);
+                if let Some(hash) = client.put_transaction(tx) {
+                    utils::print_serializable(&hash);
+                }
             }
             GET_TX => {
                 let hash = match splitted.get(1).and_then(|s| Hash::from_hex(s).ok()) {
@@ -169,7 +184,13 @@ pub fn run(mut client: Client) {
                 client.get_account_verb(id.to_owned());
             }
             GET_BLOCK => {
-                let height = match splitted.get(1).and_then(|s| s.parse::<u64>().ok()) {
+                let height = match splitted.get(1).and_then(|s| {
+                    if s.to_uppercase() == *"MAX" {
+                        Some(u64::MAX)
+                    } else {
+                        s.parse::<u64>().ok()
+                    }
+                }) {
                     Some(height) => height,
                     _ => {
                         println!("Bad input, please provide block (numeric) height");
@@ -177,6 +198,21 @@ pub fn run(mut client: Client) {
                     }
                 };
                 client.get_block_verb(height);
+            }
+            GET_TXS => {
+                let (height, mut counter) = match client.get_block(u64::MAX) {
+                    Some((blk, txs)) => (blk.data.height, txs.len()),
+                    None => {
+                        println!("There are no transactions");
+                        continue;
+                    }
+                };
+                for i in 0..height {
+                    if let Some((_, txs)) = client.get_block(i) {
+                        counter += txs.len();
+                    };
+                }
+                println!("Executed Transactions: {}", counter);
             }
             EXPORT_TXS => {
                 let _file_name = match splitted.get(1) {
@@ -203,7 +239,7 @@ pub fn run(mut client: Client) {
                 continue;
             }
             CHEATS => {
-                cheats::run(&mut client, &mut rl);
+                cheats::commons::run(&mut client, &mut rl);
                 help();
             }
             QUIT => break,